requirements.txt
# Core Data Science
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# MLOps & Experiment Tracking
mlflow==2.9.2
dvc==3.30.1
dvc-s3==3.0.1

# API & Serving
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
httpx==0.26.0

# Database
pymongo==4.6.1
python-dotenv==1.0.0

# Configuration & Serialization
pyyaml==6.0.1
joblib==1.3.2
pydantic==2.5.3
pydantic-settings==2.1.0

# Visualization & Dashboard
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
streamlit==1.29.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.92.1

# Code Quality
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Utilities
tqdm==4.66.1
click==8.1.7
requests==2.31.0
python-json-logger==2.0.7

# Jupyter (optional for notebooks)
ipykernel==6.27.1
jupyter==1.0.0
nbformat==5.9.2
